Refactoring: moving otel to exporter directory, moving utils up a directory for shareability
