+++ azureLogIngestion/adapter/processors/events.ts
@@ -2,7 +2,7 @@ import { telemetry } from "@newrelic/telemetry-sdk"
 import { Context } from "@azure/functions"
 
 import { Processor } from "./base"
-import flatten from "../utils/flatten"
+import flatten from "../../utils/flatten"
 
 export default class EventProcessor implements Processor {
     client: telemetry.events.EventClient
