+++ azureLogIngestion/exporter/opentelemetry/proto/common/v1/common.proto
