+++ azureLogIngestion/exporter/opentelemetry/proto/logs/v1/logs.proto
