+++ azureLogIngestion/exporter/opentelemetry/proto/metrics/v1/metrics.proto
