+++ azureLogIngestion/exporter/opentelemetry/proto/resource/v1/resource.proto
