+++ azureLogIngestion/exporter/opentelemetry/proto/trace/v1/trace_config.proto
