+++ azureLogIngestion/utils/camelcase.ts
