+++ azureLogIngestion/utils/flatten.test.ts
