+++ azureLogIngestion/utils/flatten.ts
