+++ azureLogIngestion/utils/mapper.test.ts
