+++ azureLogIngestion/utils/mapper.ts
