+++ azureLogIngestion/utils/normalize.ts
