+++ azureLogIngestion/utils/resource.test.ts
