+++ azureLogIngestion/utils/resource.ts
